requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
apify-client==1.0.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
boto3==1.26.102
botocore==1.29.102
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chromadb==0.3.14
click==8.1.3
clickhouse-connect==0.5.17
dataclasses-json==0.5.7
duckdb==0.7.1
fastapi==0.95.0
filelock==3.10.7
frozenlist==1.3.3
geojson==2.5.0
gevent==22.10.2
gptcache==0.1.19
greenlet==2.0.1
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
langchain==0.0.142
llama-index==0.5.23.post1
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
playwright==1.32.1
posthog==2.4.0
pydantic==1.10.7
pyee==9.0.4
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
zstandard==0.20.0
